Add Layout component and styles to main pages. Add google font to tailwindcss config.
